Assess EM model results and assign cells as damaged or not

This function accepts the EM results for a particular cell type, as well as the UMI and nuclear fraction thresholds, and assigns cells as "cell" or "damaged_cell" This is a helper function called by identify_damaged_cells and was not intended for more general use.

assess_EM(em, umi_thresh, nf_thresh)

Arguments

em
Mclust, result of running EM on the log10(UMI counts) and the nuclear fraction to asses whether there is likely to be two cell populations; cells and damaged cells, or just cells

umi_thresh
numeric, percentage of the cell disitribution UMI counts which the damaged_cell distribution must be below in order to be accepted as damaged cells. For example if the mean UMI count of the distribution fit to the cell population is 1,000 and umi_thresh is set to 30, the damaged_cell distribution mean must be less than 700 to be classified as damaged_cells, otherwise all barcodes will be returned as cells

nf_thresh
numeric, the minimum difference in the nuclear fraction score required between the two distributions (cells and damaged_cells) for the damaged_cell distribution to be accepted as real

Value

data frame, with three columns containing the log10(UMI counts), nuclear fraction score and the assigned cell status; "cell" or "damaged_cell"