Get transcript ranges

This function accepts an annotation as input and returns a GrangesList containing; GRanges defining merged exons and introns. Note that the function was written as a simple helper function to be called dropletQC::nuclear_fraction_annotation() and isn't intended for more general use.

get_transcript_ranges(
  input_annotation,
  input_annotation_format,
  ntiles,
  verbose_output
)

Arguments

input_annotation
character, should be a character vector pointing to the annotation file

input_annotation_format
character. Can be one of "auto", "gff3" or "gtf". This is passed to the 'format' argument of GenomicFeatures::makeTxDbFromGFF(). This should generally just be left as "auto"

ntiles
integer, the number of blocks that exons and introns should be grouped into

verbose_output
logical, whether to print messages

Value

GRangesList. Contains GRanges defining merged exons and introns grouped into ~ntiles blocks. When defining intronic regions, any regions that overlap exons are removed. Strand is ignored when merging all intervals.