Determine read overlap with exon and intron intervals

A function that accepts four inputs;

A GRanges defining exon and intron ranges

A reference to a BAM file

A vector of cell barcodes

A character vector describing the BAM tag which contains the cell barcode

and returns an integer twice the length of the input cell barcode vector containing; c(intron counts, exon counts), in the same order as the input barcodes. Note that the function was written as a simple helper function to be called DropletQC::nuclear_fraction_annotation() and isn't intended for more general use.

intron_exon_overlap(block, bam_file_ref, cell_barcodes, cb_tag)

Arguments

block
GRanges, contains merged exon and intron intervals

bam_file_ref
character, a reference to the BAM file to import reads from

cell_barcodes
character, a vector of cell barcodes matching the format in bam_file_ref

cb_tag
character, defines the BAM tags which contains the cell barcode e.g. "CB"

Value

integer. This function returns a vector of integers twice the length of the input cell barcode vector containing; c(intron counts, exon counts), in the same order as the input barcodes. If a cell barcode was not detected in the range, NA_integer_ is returned.