Parse cell barcode and region type tags in input region of provided BAM file

This function accepts three inputs; a GRanges defining a genomic interval, a reference to a BAM file and a vector of cell barcodes. It then uses uses Rsamtools::scanBam() to import the provided alignment tags which contain the cell barcode and region type (exonic or intronic). The tags are then parsed and returned as a data frame with two integer columns; exon and intron.

parse_bam_tags(interval, bam, bc, CB_tag, RE_tag, EXON_tag, INTRON_tag)

Arguments

interval
a single GRanges defining a genomic interval to import reads from

bam
a reference to a BAM file to pull reads from (Rsamtools::BamFile())

bc
a character vector of cell barcodes of interest

CB_tag
character, the BAM tag containing the cell barcode sequence

RE_tag
character, the BAM tag containing the region type

EXON_tag
character, the character string that defines a read as exonic

INTRON_tag
character, the character string that defines a read as intronic

Value

data.frame returns a data frame with two integer columns; "exon" - the number of reads that map to exons in the requested genomic interval for each cell barcode and "intron" - the number of reads that map to introns, where the rows are in the same order as the input character vector of cell barcodes. If there are no reads in the interval requested, the function will return NULL. This function was written as a helper function to be called internally by DropletQC::nuclear_fraction_tags and isn't intended for more general use.